[fix] 에러 메세지 수정, 권한 확인 로직 보강

+++ src/domain/comment/service/comment-command.ts
@@ -2,10 +2,14 @@ import { BusinessException } from '../../../shared/exception/business-exception/
 import { BusinessExceptionType } from '../../../shared/exception/business-exception/exception-info';
 import { TechnicalExceptionType } from '../../../shared/exception/technical-exception/exception-info';
 import { isTechnicalException } from '../../../shared/exception/technical-exception/technical-exception';
+import { IUnitOfWork } from '../../../shared/interface/i-unit-of-work';
 import { CommentEntity } from '../comment-entity';
 import { ICommentCommandRepo } from '../interface/i-comment-command-repo';
 
-export const createCommentCommandService = (commentCommandRepo: ICommentCommandRepo) => {
+export const createCommentCommandService = (
+  uow: IUnitOfWork,
+  commentCommandRepo: ICommentCommandRepo,
+) => {
   const createComment = async (
     userId: string,
     args: { content: string; resourceId: string; resourceType: 'COMPLAINT' | 'NOTICE' },
@@ -23,18 +27,26 @@ export const createCommentCommandService = (commentCommandRepo: ICommentCommandR
       if (isTechnicalException(err)) {
         if (err.type === TechnicalExceptionType.FOREIGN_KEY_VIOLATION) {
           throw BusinessException({
-            type: BusinessExceptionType.REQ_INFO_INVALID_PLEASE_RETRY,
+            type: BusinessExceptionType.FAIL_SAVE_COMMENT,
           });
         }
       }
+      throw err;
     }
   };
-  const updateComment = async (commentId: string, args: { content: string }) => {
+  const updateComment = async (userId: string, commentId: string, args: { content: string }) => {
     try {
       const beforeContext = await commentCommandRepo.findById(commentId);
 
       if (!beforeContext) {
-        throw null;
+        throw BusinessException({
+          type: BusinessExceptionType.REQ_INFO_INVALID,
+        });
+      }
+      if (beforeContext.userId !== userId) {
+        throw BusinessException({
+          type: BusinessExceptionType.FORBIDDEN,
+        });
       }
 
       const entity = CommentEntity.update(beforeContext, { content: args.content });
@@ -43,15 +55,29 @@ export const createCommentCommandService = (commentCommandRepo: ICommentCommandR
       if (isTechnicalException(err)) {
         if (err.type === TechnicalExceptionType.RECORD_NOT_FOUND) {
           throw BusinessException({
-            type: BusinessExceptionType.REQ_INFO_INVALID_PLEASE_RETRY,
+            type: BusinessExceptionType.FAIL_SAVE_COMMENT,
           });
         }
       }
+      throw err;
     }
   };
 
-  const deleteComment = async (commentId: string) => {
+  const deleteComment = async (userId: string, commentId: string) => {
     try {
+      const beforeContext = await commentCommandRepo.findById(commentId);
+
+      if (!beforeContext) {
+        throw BusinessException({
+          type: BusinessExceptionType.REQ_INFO_INVALID,
+        });
+      }
+      if (beforeContext.userId !== userId) {
+        throw BusinessException({
+          type: BusinessExceptionType.FORBIDDEN,
+        });
+      }
+
       await commentCommandRepo.delete(commentId);
     } catch (err) {
       if (isTechnicalException(err)) {
@@ -61,6 +87,7 @@ export const createCommentCommandService = (commentCommandRepo: ICommentCommandR
           });
         }
       }
+      throw err;
     }
   };
